Relocated the `GT4PyGlobalSelfCopyElimination` transformation into its own file.
