+++ src/gt4py/next/program_processors/runners/dace/transformations/self_copy_elimination.py
@@ -1,0 +1,149 @@
+# GT4Py - GridTools Framework
+#
+# Copyright (c) 2014-2024, ETH Zurich
+# All rights reserved.
+#
+# Please, refer to the LICENSE file in the root directory.
+# SPDX-License-Identifier: BSD-3-Clause
+
+from typing import Any
+
+import dace
+from dace import properties as dace_properties, transformation as dace_transformation
+from dace.sdfg import nodes as dace_nodes
+
+from gt4py.next.program_processors.runners.dace import transformations as gtx_transformations
+
+
+@dace_properties.make_properties
+class GT4PyGlobalSelfCopyElimination(dace_transformation.SingleStateTransformation):
+    """Remove global self copy.
+
+    This transformation matches the following case `(G) -> (T) -> (G)`, i.e. `G`
+    is read from and written too at the same time, however, in between is `T`
+    used as a buffer. In the example above `G` is a global memory and `T` is a
+    temporary. This situation is generated by the lowering if the data node is
+    not needed (because the computation on it is only conditional).
+
+    In case `G` refers to global memory rule 3 of ADR-18 guarantees that we can
+    only have a point wise dependency of the output on the input.
+    This transformation will remove the write into `G`, i.e. we thus only have
+    `(G) -> (T)`. The read of `G` and the definition of `T`, will only be removed
+    if `T` is not used downstream. If it is used `T` will be maintained.
+    """
+
+    node_read_g = dace_transformation.PatternNode(dace_nodes.AccessNode)
+    node_tmp = dace_transformation.transformation.PatternNode(dace_nodes.AccessNode)
+    node_write_g = dace_transformation.PatternNode(dace_nodes.AccessNode)
+
+    def __init__(
+        self,
+        *args: Any,
+        **kwargs: Any,
+    ) -> None:
+        super().__init__(*args, **kwargs)
+
+    @classmethod
+    def expressions(cls) -> Any:
+        return [dace.sdfg.utils.node_path_graph(cls.node_read_g, cls.node_tmp, cls.node_write_g)]
+
+    def can_be_applied(
+        self,
+        graph: dace.SDFGState | dace.SDFG,
+        expr_index: int,
+        sdfg: dace.SDFG,
+        permissive: bool = False,
+    ) -> bool:
+        read_g = self.node_read_g
+        write_g = self.node_write_g
+        tmp_node = self.node_tmp
+        g_desc = read_g.desc(sdfg)
+        tmp_desc = tmp_node.desc(sdfg)
+
+        # NOTE: We do not check if `G` is read downstream.
+        if read_g.data != write_g.data:
+            return False
+        if g_desc.transient:
+            return False
+        if not tmp_desc.transient:
+            return False
+        if graph.in_degree(read_g) != 0:
+            return False
+        if graph.out_degree(read_g) != 1:
+            return False
+        if graph.degree(tmp_node) != 2:
+            return False
+        if graph.in_degree(write_g) != 1:
+            return False
+        if graph.out_degree(write_g) != 0:
+            return False
+        if graph.scope_dict()[read_g] is not None:
+            return False
+
+        return True
+
+    def _is_read_downstream(
+        self,
+        start_state: dace.SDFGState,
+        sdfg: dace.SDFG,
+        data_to_look: str,
+    ) -> bool:
+        """Scans for reads to `data_to_look`.
+
+        The function will go through states that are reachable from `start_state`
+        (including) and test if there is a read to the data container `data_to_look`.
+        It will return `True` the first time it finds such a node.
+        It is important that the matched nodes, i.e. `self.node_{read_g, write_g, tmp}`
+        are ignored.
+
+        Args:
+            start_state: The state where the scanning starts.
+            sdfg: The SDFG on which we operate.
+            data_to_look: The data that we want to look for.
+
+        Todo:
+            Port this function to use DaCe pass pipeline.
+        """
+        read_g: dace_nodes.AccessNode = self.node_read_g
+        write_g: dace_nodes.AccessNode = self.node_write_g
+        tmp_node: dace_nodes.AccessNode = self.node_tmp
+
+        # TODO(phimuell): Run the `StateReachability` pass in a pipeline and use
+        #  the `_pipeline_results` member to access the data.
+        return gtx_transformations.utils.is_accessed_downstream(
+            start_state=start_state,
+            sdfg=sdfg,
+            reachable_states=None,
+            data_to_look=data_to_look,
+            nodes_to_ignore={read_g, write_g, tmp_node},
+        )
+
+    def apply(
+        self,
+        graph: dace.SDFGState | dace.SDFG,
+        sdfg: dace.SDFG,
+    ) -> None:
+        read_g: dace_nodes.AccessNode = self.node_read_g
+        write_g: dace_nodes.AccessNode = self.node_write_g
+        tmp_node: dace_nodes.AccessNode = self.node_tmp
+
+        # We first check if `T`, the intermediate is not used downstream. In this
+        #  case we can remove the read to `G` and `T` itself from the SDFG.
+        #  We have to do this check before, because the matching is not fully stable.
+        is_tmp_used_downstream = self._is_read_downstream(
+            start_state=graph, sdfg=sdfg, data_to_look=tmp_node.data
+        )
+
+        # The write to `G` can always be removed.
+        graph.remove_node(write_g)
+
+        # Also remove the read to `G` and `T` from the SDFG if possible.
+        if not is_tmp_used_downstream:
+            graph.remove_node(read_g)
+            graph.remove_node(tmp_node)
+            # It could still be used in a parallel branch.
+            try:
+                sdfg.remove_data(tmp_node.data, validate=True)
+            except ValueError as e:
+                if not str(e).startswith(f"Cannot remove data descriptor {tmp_node.data}:"):
+                    raise
